Refactor tests: replace Closure::bind with Closure::call for improved readability and maintainability
